Odoo 12 and Python3 port
* [ADD] Check route to have a response on working code

* [FIX] timesheet create on Odoo 12.0 models

* [IMP] Token generation based on python3

* [IMP] Create timesheet for specific user

* [REF] Refactoring user information

Odoo crypt password now so every time we try to get it, we have an empty string. With old code every user password would have been overwritten. With this refactoring we use an admin/manager account to write data for other users but force on creation the right user reference.

* [ADD] README and sample setting file

Co-authored-by: Apruzzese Francesco <cescoap@gmail.com>